requirements.txt
multipledispatch==0.6.0
numpy==1.19.5
Pillow==8.2.0
PyOpenGL==3.1.5
PyOpenGL-accelerate==3.1.5
six==1.15.0
